+++ packages/cinterop/src/jvm/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -744,6 +744,10 @@ actual object RealmInterop {
         realmc.sync_set_error_handler(syncConfig.cptr(), errorHandler)
     }
 
+    actual fun realm_sync_session_get(realm: NativePointer): NativePointer {
+        return LongPointerWrapper(realmc.realm_sync_session_get(realm.cptr()))
+    }
+
     actual fun realm_sync_session_wait_for_download_completion(
         syncSession: NativePointer,
         callback: SyncSessionTransferCompletionCallback
@@ -793,6 +797,31 @@ actual object RealmInterop {
 
     actual fun realm_app_credentials_new_email_password(username: String, password: String): NativePointer {
         return LongPointerWrapper(realmc.realm_app_credentials_new_email_password(username, password))
+    }
+
+    actual fun realm_app_credentials_new_api_key(key: String): NativePointer {
+        return LongPointerWrapper(realmc.realm_app_credentials_new_user_api_key(key))
+    }
+
+    actual fun realm_app_credentials_new_apple(idToken: String): NativePointer {
+        return LongPointerWrapper(realmc.realm_app_credentials_new_apple(idToken))
+    }
+
+    actual fun realm_app_credentials_new_facebook(accessToken: String): NativePointer {
+        return LongPointerWrapper(realmc.realm_app_credentials_new_facebook(accessToken))
+    }
+
+    actual fun realm_app_credentials_new_google_id_token(idToken: String): NativePointer {
+        return LongPointerWrapper(realmc.realm_app_credentials_new_google(idToken))
+    }
+
+    actual fun realm_app_credentials_new_google_auth_code(authCode: String): NativePointer {
+        TODO("See ttps://github.com/realm/realm-core/issues/5347")
+        // return LongPointerWrapper(realmc.realm_app_credentials_new_google(accessCode))
+    }
+
+    actual fun realm_app_credentials_new_jwt(jwtToken: String): NativePointer {
+        return LongPointerWrapper(realmc.realm_app_credentials_new_jwt(jwtToken))
     }
 
     actual fun realm_auth_credentials_get_provider(credentials: NativePointer): AuthProvider {