+++ test/sync/src/androidTest/kotlin/io/realm/test/mongodb/shared/CredentialsTests.kt
@@ -18,8 +18,10 @@ package io.realm.test.mongodb.shared
 
 import io.realm.internal.platform.runBlocking
 import io.realm.mongodb.App
+import io.realm.mongodb.AppException
 import io.realm.mongodb.AuthenticationProvider
 import io.realm.mongodb.Credentials
+import io.realm.mongodb.GoogleAuthType
 import io.realm.test.mongodb.TestApp
 import io.realm.test.mongodb.asTestApp
 import io.realm.test.mongodb.createUserAndLogIn
@@ -29,6 +31,8 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
+import kotlin.test.assertTrue
+import kotlin.test.fail
 
 @Suppress("ForbiddenComment")
 class CredentialsTests {
@@ -45,140 +49,127 @@ class CredentialsTests {
     @Test
     fun allCredentials() {
         for (value in AuthenticationProvider.values()) {
-            val credentials = when (value) {
-                AuthenticationProvider.ANONYMOUS ->
-                    anonymous()
-                AuthenticationProvider.EMAIL_PASSWORD ->
-                    emailPassword()
-                else -> error("Untested credentials type: $value")
+            val credentials: Credentials = when (value) {
+                AuthenticationProvider.ANONYMOUS -> anonymous()
+                AuthenticationProvider.EMAIL_PASSWORD -> emailPassword()
+                AuthenticationProvider.API_KEY -> apiKey()
+                AuthenticationProvider.APPLE -> apple()
+                AuthenticationProvider.FACEBOOK -> facebook()
+                AuthenticationProvider.GOOGLE -> { continue /* Ignore, see below */ }
+                AuthenticationProvider.JWT -> jwt()
             }
             assertEquals(value, credentials.authenticationProvider)
         }
+
+        // Special case for Google Auth having two types
+        val googleIdToken = google_idToken()
+        assertEquals(AuthenticationProvider.GOOGLE, googleIdToken.authenticationProvider)
+        // google_authCode() // Auth code not supported correctly yet: https://github.com/realm/realm-core/issues/5347
     }
 
-    fun anonymous(): Credentials {
-        val creds = Credentials.anonymous()
+    @Test
+    fun allCredentials_emptyInputThrows() {
+        for (value in AuthenticationProvider.values()) {
+            assertFailsWith<IllegalArgumentException>("$value failed") { // No arguments should be allow
+                when (value) {
+                    AuthenticationProvider.ANONYMOUS -> { throw IllegalArgumentException("Do nothing, no arguments") }
+                    AuthenticationProvider.API_KEY -> Credentials.apiKey("")
+                    AuthenticationProvider.APPLE -> Credentials.apple("")
+                    AuthenticationProvider.EMAIL_PASSWORD -> { throw IllegalArgumentException("Test below as a special case") }
+                    AuthenticationProvider.FACEBOOK -> Credentials.facebook("")
+                    AuthenticationProvider.GOOGLE -> { throw IllegalArgumentException("Test below as a special case") }
+                    AuthenticationProvider.JWT -> Credentials.jwt("")
+                }
+            }
+        }
+
+        // Test Email/Password as a special case, due to it having two arguments
+        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("", "password") }
+        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("foo@bar.com", "") }
+
+        // Test Google as a special case as two types of Google login exists
+        // TODO https://github.com/realm/realm-core/issues/5347
+        // assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.AUTH_CODE) }
+        assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.ID_TOKEN) }
+    }
+
+    private fun anonymous(): Credentials {
+        val creds: Credentials = Credentials.anonymous()
+        // TODO https://github.com/realm/realm-core/issues/5348
         // assertTrue(creds.asJson().contains("anon-user")) // Treat the JSON as an opaque value.
         return creds
     }
 
-//    @Test
-//    fun apiKey() {
-//        val creds = Credentials.apiKey("token")
-//        assertEquals(Credentials.Provider.API_KEY, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("token")) // Treat the JSON as an opaque value.
-//    }
-//
-//    @Test
-//    fun apiKey_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.apiKey("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.apiKey(TestHelper.getNull()) }
-//    }
-//
-//    @Test
-//    fun apple() {
-//        val creds = Credentials.apple("apple-token")
-//        assertEquals(Credentials.Provider.APPLE, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("apple-token")) // Treat the JSON as a largely opaque value.
-//    }
-//
-//    @Test
-//    fun apple_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.apple("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.apple(TestHelper.getNull()) }
-//    }
-//
-//    @Test
-//    fun customFunction() {
-//        val mail = TestHelper.getRandomEmail()
-//        val id = 666 + TestHelper.getRandomId()
-//        val creds = mapOf(
-//            "mail" to mail,
-//            "id" to id
-//        ).let { Credentials.customFunction(Document(it)) }
-//        assertEquals(Credentials.Provider.CUSTOM_FUNCTION, creds.identityProvider)
-//        assertTrue(creds.asJson().contains(mail))
-//        assertTrue(creds.asJson().contains(id.toString()))
-//    }
-//
-//    @Test
-//    fun customFunction_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.customFunction(null) }
-//    }
+    private fun apiKey(): Credentials {
+        val creds: Credentials = Credentials.apiKey("token")
+        // TODO https://github.com/realm/realm-core/issues/5348
+        // assertTrue(creds.asJson().contains("token")) // Treat the JSON as an opaque value.
+        return creds
+    }
 
-    fun emailPassword(): Credentials {
+    private fun apple(): Credentials {
+        val creds = Credentials.apple("apple-token")
+        // TODO https://github.com/realm/realm-core/issues/5348
+        // assertTrue(creds.asJson().contains("apple-token")) // Treat the JSON as a largely opaque value.
+        return creds
+    }
+
+    // TODO See https://github.com/realm/realm-kotlin/issues/742
+    // fun customFunction(): Credentials {
+    //     val mail = TestHelper.getRandomEmail()
+    //     val id = 666 + TestHelper.getRandomId()
+    //     val creds = mapOf(
+    //         "mail" to mail,
+    //         "id" to id
+    //     ).let { Credentials.customFunction(Document(it)) }
+    //     assertEquals(Credentials.Provider.CUSTOM_FUNCTION, creds.identityProvider)
+    //     assertTrue(creds.asJson().contains(mail))
+    //     assertTrue(creds.asJson().contains(id.toString()))
+    //     return creds
+    // }
+
+    private fun emailPassword(): Credentials {
         val creds = Credentials.emailPassword("foo@bar.com", "secret")
-        // TODO Do we need exposure of the json representation? If so, then we need exposure of it
-        //  in the C-API as well
+        // TODO https://github.com/realm/realm-core/issues/5348
         // Treat the JSON as a largely opaque value.
         // assertTrue(creds.asJson().contains("foo@bar.com"))
         // assertTrue(creds.asJson().contains("secret"))
         return creds
     }
 
-    @Test
-    fun emailPassword_invalidInput() {
-        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("", "password") }
-        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("email", "") }
-        // TODO I guess we cannot workaround these being null in Kotlin, but do we need some Java tests then?
-        // assertFailsWith<IllegalArgumentException> { Credentials.emailPassword(TestHelper.getNull(), "password") }
-        // assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("email", TestHelper.getNull()) }
+    private fun facebook(): Credentials {
+        val creds = Credentials.facebook("fb-token")
+        assertEquals(AuthenticationProvider.FACEBOOK, creds.authenticationProvider)
+        // TODO https://github.com/realm/realm-core/issues/5348
+        // assertTrue(creds.asJson().contains("fb-token"))
+        return creds
     }
 
-//    @Test
-//    fun facebook() {
-//        val creds = Credentials.facebook("fb-token")
-//        assertEquals(Credentials.Provider.FACEBOOK, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("fb-token"))
-//    }
-//
-//    @Test
-//    fun facebook_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.facebook("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.facebook(TestHelper.getNull()) }
-//    }
-//
-//    @Test
-//    fun google_authCode() {
-//        val creds = Credentials.google("google-token", GoogleAuthType.AUTH_CODE)
-//        assertEquals(Credentials.Provider.GOOGLE, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("google-token"))
-//        assertTrue(creds.asJson().contains("authCode"))
-//    }
-//
-//    @Test
-//    fun google_idToken() {
-//        val creds = Credentials.google("google-token", GoogleAuthType.ID_TOKEN)
-//        assertEquals(Credentials.Provider.GOOGLE, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("google-token"))
-//        assertTrue(creds.asJson().contains("id_token"))
-//    }
-//
-//    @Test
-//    fun google_invalidInput_authCode() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.AUTH_CODE) }
-//        assertFailsWith<IllegalArgumentException> { Credentials.google(TestHelper.getNull(), GoogleAuthType.AUTH_CODE) }
-//    }
-//
-//    @Test
-//    fun google_invalidInput_idToken() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.ID_TOKEN) }
-//        assertFailsWith<IllegalArgumentException> { Credentials.google(TestHelper.getNull(), GoogleAuthType.ID_TOKEN) }
-//    }
-//
-//    @Ignore("FIXME: Awaiting ObjectStore support")
-//    @Test
-//    fun jwt() {
-//        val creds = Credentials.jwt("jwt-token")
-//        assertEquals(Credentials.Provider.JWT, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("jwt-token"))
-//    }
-//
-//    @Test
-//    fun jwt_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.jwt("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.jwt(TestHelper.getNull()) }
-//    }
+    // TODO https://github.com/realm/realm-core/issues/5347
+    // private fun google_authCode() {
+    //  val creds = Credentials.google("google-token", GoogleAuthType.AUTH_CODE)
+    //  assertEquals(AuthenticationProvider.GOOGLE, creds.authenticationProvider)
+    // // TODO https://github.com/realm/realm-core/issues/5348
+    //  assertTrue(creds.asJson().contains("google-token"))
+    //  assertTrue(creds.asJson().contains("authCode"))
+    // }
+
+    private fun google_idToken(): Credentials {
+        val creds = Credentials.google("google-token", GoogleAuthType.ID_TOKEN)
+        assertEquals(AuthenticationProvider.GOOGLE, creds.authenticationProvider)
+        // TODO https://github.com/realm/realm-core/issues/5348
+        // assertTrue(creds.asJson().contains("google-token"))
+        // assertTrue(creds.asJson().contains("id_token"))
+        return creds
+    }
+
+    private fun jwt(): Credentials {
+        val creds = Credentials.jwt("jwt-token")
+        assertEquals(AuthenticationProvider.JWT, creds.authenticationProvider)
+        // TODO https://github.com/realm/realm-core/issues/5348
+        // assertTrue(creds.asJson().contains("jwt-token"))
+        return creds
+    }
 
     @Test
     fun loginUsingCredentials() {
@@ -190,6 +181,7 @@ class CredentialsTests {
                         val user = app.login(Credentials.anonymous())
                         assertNotNull(user)
                     }
+                    AuthenticationProvider.API_KEY -> { /* Ignore, see https://github.com/realm/realm-kotlin/issues/432 */ }
 //                    AuthenticationProvider.API_KEY -> {
 //                        // Log in, create an API key, log out, log in with the key, compare users
 //                        val user: User =
@@ -199,6 +191,9 @@ class CredentialsTests {
 //                        val apiKeyUser = app.login(Credentials.apiKey(key.value!!))
 //                        assertEquals(user.id, apiKeyUser.id)
 //                    }
+
+// TODO Enable this when https://github.com/realm/realm-kotlin/issues/741 is complete.
+
 //                    Credentials.Provider.CUSTOM_FUNCTION -> {
 //                        val customFunction = mapOf(
 //                            "mail" to TestHelper.getRandomEmail(),
@@ -217,47 +212,24 @@ class CredentialsTests {
                         val user = app.createUserAndLogIn(email, password)
                         assertNotNull(user)
                     }
-
-//                    // These providers are hard to test for real since they depend on a 3rd party
-//                    // login service. Instead we attempt to login and verify that a proper exception
-//                    // is thrown. At least that should verify that correctly formatted JSON is being
-//                    // sent across the wire.
-//                    AuthenticationProvider.FACEBOOK -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.facebook("facebook-token")
-//                        )
-//                    }
-//                    AuthenticationProvider.APPLE -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.apple("apple-token")
-//                        )
-//                    }
-//                    AuthenticationProvider.GOOGLE -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.google("google-token", GoogleAuthType.AUTH_CODE)
-//                        )
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.google("google-token", GoogleAuthType.ID_TOKEN)
-//                        )
-//                    }
-//                    AuthenticationProvider.JWT -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.jwt("jwt-token")
-//                        )
-//                    }
-//                    AuthenticationProvider.UNKNOWN -> {
-//                        // Ignore
-//                    }
+                    // These providers are hard to test for real since they depend on a 3rd party
+                    // login service. Instead we attempt to login and verify that a proper exception
+                    // is thrown. At least that should verify that correctly formatted JSON is being
+                    // sent across the wire.
+                    AuthenticationProvider.FACEBOOK -> {
+                        expectInvalidSession(app, Credentials.facebook("facebook-token"))
+                    }
+                    AuthenticationProvider.APPLE -> {
+                        expectInvalidSession(app, Credentials.apple("apple-token"))
+                    }
+                    AuthenticationProvider.GOOGLE -> {
+                        // TODO https://github.com/realm/realm-core/issues/5347
+                        // expectInvalidSession(app, Credentials.google("google-token", GoogleAuthType.AUTH_CODE))
+                        expectInvalidSession(app, Credentials.google("google-token", GoogleAuthType.ID_TOKEN))
+                    }
+                    AuthenticationProvider.JWT -> {
+                        expectInvalidSession(app, Credentials.jwt("jwt-token"))
+                    }
                     else -> {
                         error("Untested provider: $provider")
                     }
@@ -266,12 +238,14 @@ class CredentialsTests {
         }
     }
 
-//    private fun expectErrorCode(app: App, expectedCode: ErrorCode, credentials: Credentials) {
-//        try {
-//            app.login(credentials)
-//            fail()
-//        } catch (error: AppException) {
-//            assertEquals(expectedCode, error.errorCode)
-//        }
-//    }
+    private fun expectInvalidSession(app: App, credentials: Credentials) {
+        try {
+            runBlocking {
+                app.login(credentials)
+            }
+            fail()
+        } catch (error: AppException) {
+            assertTrue(error.message!!.contains("authentication via"), error.message)
+        }
+    }
 }